requirements.txt
pandas>=2.0
numpy>=1.24
scipy>=1.10
scikit-learn>=1.3
statsmodels>=0.14
matplotlib>=3.7
openpyxl>=3.1
pyyaml>=6.0
requests>=2.31
tqdm>=4.66
python-dotenv>=1.0
